requirements.txt
pandas==1.3.4
plotly-express==0.4.1
psutil==5.8.0
pyspark==3.2.0
streamlit==1.2.0